move AppParamParser.h/cpp from xbmc/settings/ to xbmc/
